backend/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-core==2.23.4
numpy==1.26.4
SQLAlchemy==2.0.34
PyMySQL==1.1.1
python-dotenv==1.0.1
